fix warnings

+++ src/xtd_forms_native_wxwidgets/src/xtd/forms/native/wxwidgets/checked_list_box.cpp
@@ -16,17 +16,17 @@ void checked_list_box::begin_update(intptr_t control) {
 
 void checked_list_box::check_state(intptr_t control, size_t index, int check_state) {
   if (control == 0) return;
-  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->Check(index, check_state != 0);
+  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->Check(static_cast<int32_t>(index), check_state != 0);
 }
 
 int checked_list_box::check_state(intptr_t control, size_t index) {
   if (control == 0) return false;
-  return static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->IsChecked(index);
+  return static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->IsChecked(static_cast<int32_t>(index));
 }
 
 void checked_list_box::delete_item(intptr_t control, size_t index) {
   if (control == 0) return;
-  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->Delete(index);
+  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->Delete(static_cast<int32_t>(index));
 }
 
 void checked_list_box::end_update(intptr_t control) {
@@ -36,8 +36,8 @@ void checked_list_box::end_update(intptr_t control) {
 
 void checked_list_box::insert_item(intptr_t control, size_t index, const std::string& value, bool checked) {
   if (control == 0) return;
-  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->Insert({value.c_str(), wxMBConvUTF8()}, index);
-  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->Check(index, checked);
+  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->Insert({value.c_str(), wxMBConvUTF8()}, static_cast<int32_t>(index));
+  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->Check(static_cast<int32_t>(index), checked);
 #if defined(__WXMSW__)
   static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->GetItem(index)->SetTextColour(static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->GetForegroundColour());
 #endif
@@ -55,7 +55,7 @@ size_t checked_list_box::selected_index(intptr_t control) {
 
 void checked_list_box::selected_index(intptr_t control, size_t index) {
   if (control == 0) return;
-  return static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->SetSelection(index);
+  return static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->SetSelection(static_cast<int32_t>(index));
 }
 
 vector<size_t> checked_list_box::selected_indices(intptr_t control) {
@@ -71,6 +71,6 @@ vector<size_t> checked_list_box::selected_indices(intptr_t control) {
 
 void checked_list_box::update_item(intptr_t control, size_t index, const std::string& value, bool checked) {
   if (control == 0) return;
-  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->SetString(index, {value.c_str(), wxMBConvUTF8()});
-  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->Check(index, checked);
+  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->SetString(static_cast<int32_t>(index), {value.c_str(), wxMBConvUTF8()});
+  static_cast<wxCheckListBox*>(reinterpret_cast<control_handler*>(control)->control())->Check(static_cast<int32_t>(index), checked);
 }
